# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LQGovernance — Board Management Portal API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /legal-holds:
    get:
      operationId: listLegalHolds
      tags: [legal-holds]
      summary: List legal holds (admin only; active by default, ?all=true for released)
      responses:
        "200":
          description: Holds
          content:
            application/json:
              schema: { type: array, items: { type: object } }
    post:
      operationId: placeLegalHold
      tags: [legal-holds]
      summary: Place a legal hold on an entity or board (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceLegalHoldBody"
      responses:
        "201": { description: Hold placed }
        "400": { description: Invalid input }
  /legal-holds/{id}/release:
    post:
      operationId: releaseLegalHold
      tags: [legal-holds]
      summary: Release a legal hold (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Released }
        "404": { description: Not found }
        "409": { description: Already released }
...
components:
  schemas:
...
    PlaceLegalHoldBody:
      type: object
      required: [entityType, entityId, reason]
      properties:
        entityType:
          type: string
          enum: [board, meeting, document, vote, task]
        entityId:
          type: string
        reason:
          type: string
...